+++ lizard_measure/models.py
@@ -175,7 +175,7 @@ class WaterBody(models.Model):
         if self.area is None:
             return 'area_ident: %s (No geometry)' % self.area_ident
         else:
-            return u'%s' % self.area.name
+            return u'%s (%s)' % (self.area.name, self.area.ident)
 
 
 class MeasuringRodManager(models.Manager):
@@ -1189,8 +1189,9 @@ class Measure(models.Model):
         help_text=('Wanneer een maatregel read-only is, is hij geimporteerd '
                    'en dient hij niet met de hand gewijzigd te worden'))
 
-    # Is this different from is_KRW_measure?
-    is_indicator = models.BooleanField(default=False)
+    #
+    is_indicator = models.BooleanField(default=False,
+        help_text='focus maatregel')
 
     data_set = models.ForeignKey(DataSet,
                                  null=True,
@@ -1337,13 +1338,12 @@ class Measure(models.Model):
             id = EsfLink.id
             and other fields of esflink
         """
+
         for esflink in esflinks:
 
             esf = self.esflink_set.get(
                 pk=esflink['id']
             )
-
-            print esflink
 
             esf.is_target_esf = esflink['is_target_esf']
             esf.positive = esflink['positive']
@@ -1363,6 +1363,7 @@ class Measure(models.Model):
             positive
             negative
         """
+
         output = []
 
         if auto_create_missing_states: